requirements.txt
dash
langchain_community
langchain
langchain_openai
python-dotenv